+++ src/components/Mobile/MobileProject/MobileProject.js
@@ -5,17 +5,23 @@ import Output from "../../Editor/Output/Output";
 import MobileProjectBar from "./../MobileProjectBar/MobileProjectBar";
 
 import "./MobileProject.scss";
-import { useSelector } from "react-redux";
-import { CodeIcon, PreviewIcon } from "../../../Icons";
+import { useDispatch, useSelector } from "react-redux";
+import { CodeIcon, PreviewIcon, MenuIcon } from "../../../Icons";
 import { useTranslation } from "react-i18next";
+import Sidebar from "../../Menus/Sidebar/Sidebar";
+import { showSidebar } from "../../Editor/EditorSlice";
 
 const MobileProject = () => {
   const projectType = useSelector((state) => state.editor.project.project_type);
+  const sidebarShowing = useSelector((state) => state.editor.sidebarShowing);
   const codeRunTriggered = useSelector(
     (state) => state.editor.codeRunTriggered,
   );
   const [selectedTab, setSelectedTab] = useState(0);
   const { t } = useTranslation();
+  const dispatch = useDispatch();
+
+  const openSidebar = () => dispatch(showSidebar());
 
   useEffect(() => {
     if (codeRunTriggered) {
@@ -24,39 +30,51 @@ const MobileProject = () => {
   }, [codeRunTriggered]);
 
   return (
-    <div className="proj-container proj-editor-container proj-container--mobile">
-      <Tabs
-        forceRenderTabPanel={true}
-        selectedIndex={selectedTab}
-        onSelect={(index) => setSelectedTab(index)}
-      >
-        <TabPanel>
-          <EditorInput />
-        </TabPanel>
-        <TabPanel>
-          <Output />
-        </TabPanel>
-        <MobileProjectBar />
-        <div className="react-tabs__tab-container mobile-nav">
-          <TabList>
-            <Tab>
-              <span className="react-tabs__tab-inner">
-                <CodeIcon />
-                {t("mobile.code")}
+    <>
+      {sidebarShowing ? (
+        <Sidebar />
+      ) : (
+        <div className="proj-container proj-editor-container proj-container--mobile">
+          <Tabs
+            forceRenderTabPanel={true}
+            selectedIndex={selectedTab}
+            onSelect={(index) => setSelectedTab(index)}
+          >
+            <TabPanel>
+              <EditorInput />
+            </TabPanel>
+            <TabPanel>
+              <Output />
+            </TabPanel>
+            <MobileProjectBar />
+            <div className="react-tabs__tab-container mobile-nav">
+              <span
+                className="react-tabs__tab-inner mobile-nav__menu"
+                onClick={openSidebar}
+              >
+                <MenuIcon />
               </span>
-            </Tab>
-            <Tab>
-              <span className="react-tabs__tab-inner">
-                <PreviewIcon />
-                {projectType === "html"
-                  ? t("mobile.preview")
-                  : t("mobile.output")}
-              </span>
-            </Tab>
-          </TabList>
+              <TabList>
+                <Tab>
+                  <span className="react-tabs__tab-inner">
+                    <CodeIcon />
+                    {t("mobile.code")}
+                  </span>
+                </Tab>
+                <Tab>
+                  <span className="react-tabs__tab-inner">
+                    <PreviewIcon />
+                    {projectType === "html"
+                      ? t("mobile.preview")
+                      : t("mobile.output")}
+                  </span>
+                </Tab>
+              </TabList>
+            </div>
+          </Tabs>
         </div>
-      </Tabs>
-    </div>
+      )}
+    </>
   );
 };