+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - Mobile navigation between code and output (#615)
+- Mobie sidebar
 
 ### Changed