+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 
 - Position of run button on mobile (#615)
+- Save button styles (#633)
 
 ## [0.18.1] - 2023-08-03